Reorganise AFL application layer into domain-concern files
Merge ScoreCommands into Commands (artificial struct split removed).
Split Commands methods across player_match.go, score.go, reactions.go
by reason-to-change. commands.go now holds only the struct and constructor.
Renamed data_ops.go → dataops.go to match FFL naming.
