+++ services/afl/internal/interface/graphql/resolver.go
@@ -4,8 +4,7 @@ import "xffl/services/afl/internal/application"
 
 // Resolver is the dependency injection container for GraphQL resolvers.
 type Resolver struct {
-	Queries       *application.Queries
-	Commands      *application.Commands
-	DataOps       *application.DataOpsCommands
-	ScoreCommands *application.ScoreCommands
+	Queries  *application.Queries
+	Commands *application.Commands
+	DataOps  *application.DataOpsCommands
 }
